requirements.txt
speechmatics_flow
pyaudio
python-dotenv